demoApp/requirements.txt
adlfs==2024.7.0
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
azure-core==1.30.2
azure-datalake-store==0.0.53
azure-functions==1.20.0
azure-identity==1.17.1
azure-storage-blob==12.22.0
blinker==1.8.2
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
dnspython==2.6.1
email_validator==2.2.0
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
frozenlist==1.4.1
fsspec==2024.6.1
greenlet==3.0.3
idna==3.7
importlib_metadata==8.2.0
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
msal==1.30.0
msal-extensions==1.2.0
multidict==6.0.5
numpy==1.24.4
pandas==2.0.3
portalocker==2.10.1
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.32
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
WTForms==3.1.2
yarl==1.9.4
zipp==3.19.2
